backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
fastapi==0.115.12
h11==0.16.0
idna==3.10
lxml==5.4.0
pydantic==2.11.5
pydantic_core==2.33.2
PyMuPDF==1.26.0
python-docx==1.1.2
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
